refactor: refactored the whole search functionality and some other stuff
Refactored a lot in the search functionality to get the Application state be consistent throughout the entire application.
Also refactored the BookShelf into Books, to get the whole stuff smaler.
Added three new packages, to get the functionality that is needed.
Refactored the loader function, to get a smarter loader in search function.